% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{add_file_paths}
\alias{add_file_paths}
\title{Function to add file_paths to mapping table}
\usage{
add_file_paths(
  lookup_table,
  figure_export_path,
  table_export_path,
  dataset_export_path
)
}
\description{
Function to add file_paths to mapping table
}
\examples{
## Placeholder Example ##
}
