% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{add_source_type}
\alias{add_source_type}
\title{Function to add source_type to mapping table}
\usage{
add_source_type(lookup_table)
}
\description{
Function to add source_type to mapping table
}
\examples{
## Placeholder Example ##
}
