% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{add_yaml_header}
\alias{add_yaml_header}
\title{Function to add yaml header to .qmd file}
\usage{
add_yaml_header(
  bibliography,
  csl_style,
  format = "docx",
  include_refs = TRUE,
  include_figures = FALSE,
  dataset_export_path = NULL,
  table_export_path = NULL,
  figure_export_path = NULL
)
}
\description{
Function to add yaml header to .qmd file
}
\examples{
## Placeholder Example ##
}
