% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{apply_base_visual_style}
\alias{apply_base_visual_style}
\title{Function to apply base visual style}
\usage{
apply_base_visual_style(style_name = "dataStyle", first_neighbors = FALSE)
}
\description{
Function to apply base visual style
}
\examples{
## Placeholder Example ##
}
