% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{apply_edge_confidence}
\alias{apply_edge_confidence}
\title{Function to map STRING confidence to edge width}
\usage{
apply_edge_confidence(style_name = "dataStyle", min_width = 2, max_width = 6)
}
\description{
Function to map STRING confidence to edge width
}
\examples{
## Placeholder Example ##
}
