% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{apply_feature_policy}
\alias{apply_feature_policy}
\title{Function for determining how to deal with feature outliers}
\usage{
apply_feature_policy(
  df,
  proteins,
  nzv,
  cor,
  mad,
  policy,
  freq_thrs,
  verbose = TRUE
)
}
\description{
Function for determining how to deal with feature outliers
}
\examples{
## Placeholder Example ##
}
