% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{apply_sample_policy}
\alias{apply_sample_policy}
\title{Function for determining how to deal with sample outliers}
\usage{
apply_sample_policy(
  df,
  strong,
  moderate,
  policy,
  grouping_variable,
  sample_rm_thr,
  sample_diff_thr,
  verbose = TRUE
)
}
\description{
Function for determining how to deal with sample outliers
}
\examples{
## Placeholder Example ##
}
