% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{apply_visual_contrast}
\alias{apply_visual_contrast}
\title{Function to apply specific visual style according to contrast}
\usage{
apply_visual_contrast(
  contrast,
  style_name = "dataStyle",
  border_colors = c("deepskyblue2", "firebrick1"),
  min_border_width = 50,
  max_border_width = 100,
  default_edge_color = "gray70",
  color_edges = FALSE,
  edge_sets = NULL,
  edge_palette = NULL,
  edge_width_highlight = 20,
  expr_col = NULL,
  width_col = NULL,
  class_col = NULL
)
}
\description{
Function to apply specific visual style according to contrast
}
\examples{
## Placeholder Example ##
}
