% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{assemble_risk}
\alias{assemble_risk}
\title{Function to assemble out-of-fold risk vector from fold results}
\usage{
assemble_risk(fold_results, df)
}
\description{
Function to assemble out-of-fold risk vector from fold results
}
\examples{
## Placeholder Example ##
}
