% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{binary_de_model}
\alias{binary_de_model}
\title{Function to compute DE analysis with limma with two variables}
\usage{
binary_de_model(
  df,
  grouping_var,
  proteins,
  covars = NULL,
  pval_cutoff = 0.05,
  LFC_cutoff = 0.25,
  treat = FALSE,
  ref_level = NULL,
  verbose = TRUE
)
}
\description{
Function to compute DE analysis with limma with two variables
}
\examples{
## Placeholder Example ##
}
