% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{build_qc_report}
\alias{build_qc_report}
\title{Function to get QC summary}
\usage{
build_qc_report(
  n_samples_original,
  n_samples_final,
  n_proteins_original,
  n_proteins_final,
  strong_outliers,
  moderate_outliers,
  samples_removed,
  proteins_removed,
  freq_threshold,
  imbalance_flag,
  sample_plots = NULL,
  feature_plots = NULL
)
}
\description{
Function to get QC summary
}
\examples{
## Placeholder Example ##
}
