% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_wrapper.R
\name{build_string_network}
\alias{build_string_network}
\title{Function to wrap protein networks pipeline}
\usage{
build_string_network(
  input_table,
  mode = c("multi", "single"),
  first_neighbors = FALSE,
  hide_singletons = TRUE,
  show_first_neighbors = FALSE,
  neighbor_min_connections = string_neighbor_min_connections,
  neighbor_score_threshold = string_neighbor_score_threshold,
  net_proteins = NULL,
  wait_for_layout_adj = TRUE
)
}
\description{
Function to wrap protein networks pipeline
}
\examples{
## Placeholder Example ##
}
