% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{chk_gp_imbalance}
\alias{chk_gp_imbalance}
\title{Function for checking group imbalance (var=variable of choice)}
\usage{
chk_gp_imbalance(
  df_unfilt,
  df_filt,
  var,
  sample_rm_thr = 30,
  sample_diff_thr = 20
)
}
\description{
Function for checking group imbalance (var=variable of choice)
}
\examples{
## Placeholder Example ##
}
