% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_protein_names}
\alias{clean_protein_names}
\title{Function to adjust protein notation}
\usage{
clean_protein_names(protein_names, strict = TRUE)
}
\description{
Function to adjust protein notation
}
\examples{
## Placeholder Example ##
}
