% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_spliting_data.R
\name{collapse_quantile_groups}
\alias{collapse_quantile_groups}
\title{Function to collapse quantile groups into other groups (e.g. 'high' vs 'low')}
\usage{
collapse_quantile_groups(data, feature, groups, new_col = NULL, verbose = TRUE)
}
\description{
Function to collapse quantile groups into other groups (e.g. 'high' vs 'low')
}
\examples{
## Placeholder Example ##
}
