% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{compare_ref_network}
\alias{compare_ref_network}
\title{Function to compare networks against reference}
\usage{
compare_ref_network(
  ptn_net_list,
  ref_ptn,
  first_neighbors = FALSE,
  neighbor_min_connections = 1,
  string_neighbor_score_threshold = 700
)
}
\description{
Function to compare networks against reference
}
\examples{
## Placeholder Example ##
}
