% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_data.R
\name{create_all_subsets}
\alias{create_all_subsets}
\title{Function wrapper to generate subsets with many proteins and variable groups}
\usage{
create_all_subsets(
  df_normal,
  df_disease,
  df_apply,
  protein_list = NULL,
  variable_list = NULL,
  var_labs = NULL,
  normal_pattern = "CD34",
  ref_patterns = c("WT", "wt", "no"),
  exclude_NA = TRUE,
  print_summary = TRUE
)
}
\description{
Function wrapper to generate subsets with many proteins and variable groups
}
\examples{
## Placeholder Example ##
}
