% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{create_base_network}
\alias{create_base_network}
\title{Function to create base network}
\usage{
create_base_network(
  nodes,
  edges,
  title = "Protein Network",
  collection = "Proteomics",
  hide_singletons = TRUE
)
}
\description{
Function to create base network
}
\examples{
## Placeholder Example ##
}
