% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{create_bxplot}
\alias{create_bxplot}
\title{Function to create boxplots}
\usage{
create_bxplot(
  df,
  continuous_var,
  discrete_var,
  fill_var = NULL,
  padj = "BH",
  colors = NULL,
  title = NULL,
  show_pairwise = FALSE,
  show_median_line = TRUE,
  label_y_test_pos_pct = 0.1,
  angle_x_var = 0,
  legend_bxplt = FALSE,
  legend_jitter = FALSE,
  fill_label = NULL,
  ylab = NULL,
  xlab = NULL
)
}
\description{
Function to create boxplots
}
\examples{
## Placeholder Example ##
}
