% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{create_zotero_collection}
\alias{create_zotero_collection}
\title{Function to create Zotero collection}
\usage{
create_zotero_collection(
  name,
  user_id,
  api_key,
  main_cache_path,
  zotero_version = "3",
  parent_key = NULL
)
}
\description{
Function to create Zotero collection
}
\examples{
## Placeholder Example ##
}
