% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_utilities.R
\name{define_csl_style}
\alias{define_csl_style}
\title{Function wrapper to define CSL style}
\usage{
define_csl_style(journal, cache_path)
}
\description{
Function wrapper to define CSL style
}
\examples{
## Placeholder Example ##
}
