% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{detect_feature_outliers}
\alias{detect_feature_outliers}
\title{Function to detect feature outliers}
\usage{
detect_feature_outliers(
  df,
  proteins,
  check_cor,
  check_mad,
  cor_cutoff,
  ptn_z_thrs,
  freq_thrs,
  feature_outlier_policy = "keep_all"
)
}
\description{
Function to detect feature outliers
}
\examples{
## Placeholder Example ##
}
