% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{detect_sample_outliers}
\alias{detect_sample_outliers}
\title{Function for sample outlier detection}
\usage{
detect_sample_outliers(df, proteins)
}
\description{
Function for sample outlier detection
}
\examples{
## Placeholder Example ##
}
