% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_spliting_data.R
\name{determine_optimal_quantile}
\alias{determine_optimal_quantile}
\title{Function to automate best quantile split determination}
\usage{
determine_optimal_quantile(
  data,
  feature,
  time_var,
  event_var,
  q_range = 2:6,
  padj_method = "BH"
)
}
\description{
Function to automate best quantile split determination
}
\examples{
## Placeholder Example ##
}
