% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{enrichr_path}
\alias{enrichr_path}
\title{Function for EnrichR analysis}
\usage{
enrichr_path(
  proteins,
  libs,
  enrichr_filt_ptn = FALSE,
  min_ptn_enrichr_path = 3,
  enrichr_filt_path = FALSE,
  min_path_enrichr_ptn = 3,
  out_rm = TRUE,
  drop_paths = TRUE,
  path_number = 20
)
}
\description{
Function for EnrichR analysis
}
\examples{
## Placeholder Example ##
}
