% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_bxplot_tests}
\alias{export_bxplot_tests}
\title{Function to export tables to excel spreadsheets}
\usage{
export_bxplot_tests(results, filename, filepath, tab_title)
}
\description{
Function to export tables to excel spreadsheets
}
\examples{
## Placeholder Example ##
}
