% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_bxplots_pdf}
\alias{export_bxplots_pdf}
\title{Function to export boxplots}
\usage{
export_bxplots_pdf(results, file = filename, ncol = 3, nrow = 2)
}
\description{
Function to export boxplots
}
\examples{
## Placeholder Example ##
}
