% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_tabs}
\alias{export_tabs}
\title{Function to Export tables to excel}
\usage{
export_tabs(df_list, filename, filepath, tab_title)
}
\description{
Function to Export tables to excel
}
\examples{
## Placeholder Example ##
}
