% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{extract_float_order}
\alias{extract_float_order}
\title{Function to extract keys from text in order of appearance}
\usage{
extract_float_order(manuscript_path, lookup_table)
}
\description{
Function to extract keys from text in order of appearance
}
\examples{
## Placeholder Example ##
}
