% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{extract_ids}
\alias{extract_ids}
\title{Function to extract and validate identifiers from text}
\usage{
extract_ids(text, input_pattern, output_pattern, known_ids = character(0))
}
\description{
Function to extract and validate identifiers from text
}
\examples{
## Placeholder Example ##
}
