% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{filter_lookup_table}
\alias{filter_lookup_table}
\title{Function to select file types in lookup table to export according to rendering mode}
\usage{
filter_lookup_table(lookup_table, mode = c("draft", "review", "submission"))
}
\description{
Function to select file types in lookup table to export according to rendering mode
}
\examples{
## Placeholder Example ##
}
