% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{get_best_seed}
\alias{get_best_seed}
\title{Function to test which seed is best to separate train and test data}
\usage{
get_best_seed(train_data, test_data, event_var, max_seed = 2026)
}
\description{
Function to test which seed is best to separate train and test data
}
\examples{
## Placeholder Example ##
}
