% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_data_mining.R
\name{get_cbio_data}
\alias{get_cbio_data}
\title{Function to download, decompress and load studies from cBioPortal}
\usage{
get_cbio_data(
  study_id,
  main_cache_path,
  cache = TRUE,
  verbose = TRUE,
  force = FALSE
)
}
\description{
Function to download, decompress and load studies from cBioPortal
}
\examples{
## Placeholder Example ##
}
