% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{get_comb_sig_tables}
\alias{get_comb_sig_tables}
\title{Function to merge descriptive tables of significant variables}
\usage{
get_comb_sig_tables(
  table_list,
  grouping_variable = "group",
  exclude_sig = NULL,
  label_vars = NULL,
  dichotomous_vars = NULL,
  seed = 123
)
}
\description{
Function to merge descriptive tables of significant variables
}
\examples{
## Placeholder Example ##
}
