% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{get_descr_tab}
\alias{get_descr_tab}
\title{Function to make descriptive table with gtsummary}
\usage{
get_descr_tab(
  df,
  grouping_variable,
  vars,
  skip_pval = -9999,
  vars_to_test = NULL,
  tests = NULL,
  args = NULL,
  skipped_vars = NULL,
  exclude_sig = NULL,
  sig_type = NULL,
  sig_value = NULL,
  sig_labels = NULL,
  seed = 123
)
}
\description{
Function to make descriptive table with gtsummary
}
\examples{
## Placeholder Example ##
}
