% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{get_feature_dataset}
\alias{get_feature_dataset}
\title{Function to get protein dataset}
\usage{
get_feature_dataset(data, n_features, id_column, start_column)
}
\description{
Function to get protein dataset
}
\examples{
## Placeholder Example ##
}
