% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mycolors}
\alias{get_mycolors}
\title{Function to get colors from pre-defined sets or used-defined sets}
\usage{
get_mycolors(
  n_colors = NULL,
  mode = c("list", "vector"),
  color_list = NULL,
  color_vector = NULL
)
}
\description{
Function to get colors from pre-defined sets or used-defined sets
}
\examples{
## Placeholder Example ##
}
