% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_data.R
\name{get_ptn_groups}
\alias{get_ptn_groups}
\title{Function to make protein groups}
\usage{
get_ptn_groups(
  df_normal,
  protein,
  df_disease = NULL,
  df_apply = NULL,
  normal_pattern = "CD34"
)
}
\description{
Function to make protein groups
}
\examples{
## Placeholder Example ##
}
