% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_spliting_data.R
\name{get_quantile_groups}
\alias{get_quantile_groups}
\title{Function to get several quantile groups for many proteins}
\usage{
get_quantile_groups(
  data,
  features,
  n_quantiles = NULL,
  data_normal = NULL,
  mode = c("quantile", "normal_based", "both"),
  normal_labels = c("Below Normal", "Lower Normal", "Upper Normal", "Above Normal")
)
}
\description{
Function to get several quantile groups for many proteins
}
\examples{
## Placeholder Example ##
}
