% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{get_rawtable_values}
\alias{get_rawtable_values}
\title{Function to get raw values from raw datatable}
\usage{
get_rawtable_values(data, features)
}
\description{
Function to get raw values from raw datatable
}
\examples{
## Placeholder Example ##
}
