% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{get_reference_table}
\alias{get_reference_table}
\title{Function to extract and validate all IDs}
\usage{
get_reference_table(
  text,
  collection_name,
  main_cache_path,
  crossref_email = NULL,
  pmid_field_pattern = "\\\\(\\\\s*PMID\\\\s*:[^)]*\\\\)",
  pmid_clean_pattern = "\\\\d{5,9}",
  doi_field_pattern = "\\\\(\\\\s*DOI\\\\s*:[^()]*(?:\\\\([^()]*\\\\)[^()]*)*\\\\)",
  doi_clean_pattern =
    "10\\\\.\\\\d{4,9}/(?:[^\\\\s,:;()]*(?:\\\\([^()]*\\\\))?)+(?=\\\\.?(?:\\\\s*[,:;)]|\\\\)$))"
)
}
\description{
Function to extract and validate all IDs
}
\examples{
## Placeholder Example ##
}
