% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{get_single_ptn_network}
\alias{get_single_ptn_network}
\title{Function to get a single protein network}
\usage{
get_single_ptn_network(
  protein_name,
  first_neighbors = TRUE,
  neighbor_min_connections = 1,
  neighbor_score_threshold = string_neighbor_score_threshold,
  stringdb_version = string_version,
  stringdb_score_threshold = string_score_threshold,
  stringdb_cache_path = string_cache_path,
  stringdb_species = string_species
)
}
\description{
Function to get a single protein network
}
\examples{
## Placeholder Example ##
}
