% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{get_stringdb_data}
\alias{get_stringdb_data}
\title{Function to extract networks with StringDB}
\usage{
get_stringdb_data(
  input_table,
  first_neighbors = FALSE,
  neighbor_min_connections = string_neighbor_min_connections,
  neighbor_score_threshold = string_neighbor_score_threshold,
  stringdb_version = string_version,
  stringdb_score_threshold = string_score_threshold,
  stringdb_cache_path = string_cache_path,
  stringdb_species = string_species
)
}
\description{
Function to extract networks with StringDB
}
\examples{
## Placeholder Example ##
}
