% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_data_mining.R
\name{get_study_data}
\alias{get_study_data}
\title{Function to extract relevant data from cBioPortal and create a single dataset}
\usage{
get_study_data(
  data,
  study_id,
  main_cache_path,
  assay_names = NULL,
  assay_labels = NULL,
  patient_id_col = "PATIENT_ID",
  sample_id_col = "SAMPLE_ID",
  verbose = TRUE,
  cache = TRUE,
  force = FALSE
)
}
\description{
Function to extract relevant data from cBioPortal and create a single dataset
}
\examples{
## Placeholder Example ##
}
