% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_data.R
\name{get_subsets}
\alias{get_subsets}
\title{Function to create patient subsets}
\usage{
get_subsets(df, group_lists, print_summary = TRUE, exclude_NA = TRUE)
}
\description{
Function to create patient subsets
}
\examples{
## Placeholder Example ##
}
