% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{get_upset_ptns}
\alias{get_upset_ptns}
\title{Function to extract protein names from upset plot}
\usage{
get_upset_ptns(comb_mat)
}
\description{
Function to extract protein names from upset plot
}
\examples{
## Placeholder Example ##
}
