% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_data.R
\name{get_var_groups}
\alias{get_var_groups}
\title{Function to make variable groups}
\usage{
get_var_groups(
  df,
  var,
  var_labs = NULL,
  ref_patterns = c("WT", "wt", "no"),
  ref_suffix = "neg",
  alt_suffix = "pos",
  missing_suffix = "missing",
  verbose = TRUE
)
}
\description{
Function to make variable groups
}
\examples{
## Placeholder Example ##
}
