% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{get_zotero_bib}
\alias{get_zotero_bib}
\title{Function to get references.bib from Zotero}
\usage{
get_zotero_bib(collection_name, collection_key, main_cache_path)
}
\description{
Function to get references.bib from Zotero
}
\examples{
## Placeholder Example ##
}
