% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{handle_all_sample_outliers}
\alias{handle_all_sample_outliers}
\title{Function for interaction after sample outlier detection}
\usage{
handle_all_sample_outliers(
  strong_out_sample = character(0),
  moderate_out_sample = character(0),
  remove_strong_only = FALSE,
  remove_all = FALSE,
  keep_all = FALSE,
  allow_interactive_override = TRUE,
  verbose = TRUE
)
}
\description{
Function for interaction after sample outlier detection
}
\examples{
## Placeholder Example ##
}
