% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{highlight_edge_sets}
\alias{highlight_edge_sets}
\title{Function to color edges from specific networks}
\usage{
highlight_edge_sets(net_proteins)
}
\description{
Function to color edges from specific networks
}
\examples{
## Placeholder Example ##
}
