% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_analysis.R
\name{highlight_hubs}
\alias{highlight_hubs}
\title{Function to ajust nodes according to protein connections (used inside apply_base_visual_style)}
\usage{
highlight_hubs(
  style_name = "dataStyle",
  min_size = 300,
  max_size = 500,
  top_percent = 0.1
)
}
\description{
Function to ajust nodes according to protein connections (used inside apply_base_visual_style)
}
\examples{
## Placeholder Example ##
}
