% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{interleave_results}
\alias{interleave_results}
\title{Function for flexible interleaving multi-plot results (customize order of tables and plots with optional protein/group naming)}
\usage{
interleave_results(
  list_of_lists,
  protein_names = NULL,
  group_names = NULL,
  include_protein = TRUE,
  include_group = TRUE,
  labels = NULL,
  suffixes = NULL
)
}
\description{
Function for flexible interleaving multi-plot results (customize order of tables and plots with optional protein/group naming)
}
\examples{
## Placeholder Example ##
}
