% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{join_citekeys}
\alias{join_citekeys}
\title{Function to join citekeys properly}
\usage{
join_citekeys(table, citekeys)
}
\description{
Function to join citekeys properly
}
\examples{
## Placeholder Example ##
}
