% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_utilities.R
\name{load_csl_registry}
\alias{load_csl_registry}
\title{Function to register CSL entries}
\usage{
load_csl_registry()
}
\description{
Function to register CSL entries
}
\examples{
## Placeholder Example ##
}
