% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{make_float_chunks}
\alias{make_float_chunks}
\title{Function to generate float chunks to embed within text}
\usage{
make_float_chunks(lookup_table)
}
\description{
Function to generate float chunks to embed within text
}
\examples{
## Placeholder Example ##
}
