% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{manuscript_export}
\alias{manuscript_export}
\title{Function to wrap manuscript exporting}
\usage{
manuscript_export(
  manuscript_path,
  ref_bib_path,
  csl_cache_path,
  lookup_table,
  dataset_export_path = NULL,
  table_export_path = NULL,
  figure_export_path = NULL,
  output_path = NULL,
  ref_header = "**REFERENCES**",
  figure_table_header = "FIGURES AND TABLES",
  sup_naming = "Supplemental",
  journal_ref = "ajh",
  manuscript_name = "manuscript",
  format = "docx",
  mode = c("draft", "review", "submission")
)
}
\description{
Function to wrap manuscript exporting
}
\examples{
## Placeholder Example ##
}
