% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{membership_to_list}
\alias{membership_to_list}
\title{Function used to transform data input into a list}
\usage{
membership_to_list(data, membership_col, yes_value = "yes")
}
\description{
Function used to transform data input into a list
}
\examples{
## Placeholder Example ##
}
