% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{mv_variable_selection}
\alias{mv_variable_selection}
\title{Function to perform forward selection in MV models}
\usage{
mv_variable_selection(
  data,
  candidate_vars,
  mv_vars_pool,
  time_var,
  event_var,
  grouping_variable,
  max_predictors,
  meta_table,
  min_epv = 10,
  cluster_var_p_cutoff = 0.05,
  lrt_p_cutoff = 0.05,
  block_vars = TRUE,
  selection_strategy = c("strict", "two_phase", "cie"),
  cie_threshold = 0.1,
  cie_final_cleanup = FALSE,
  all_levels_sig = FALSE,
  forced_selected = NULL,
  verbose = TRUE
)
}
\description{
Function to perform forward selection in MV models
}
\examples{
## Placeholder Example ##
}
