% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{ordered_de_model}
\alias{ordered_de_model}
\title{Function to compute DE analysis with limma with three or more variables (ordered model)}
\usage{
ordered_de_model(
  df,
  grouping_var,
  proteins,
  covars = NULL,
  pval_cutoff = 0.05,
  LFC_cutoff = 0.25,
  verbose = TRUE
)
}
\description{
Function to compute DE analysis with limma with three or more variables (ordered model)
}
\examples{
## Placeholder Example ##
}
