% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{path_heatmap}
\alias{path_heatmap}
\title{Function to plot Heatmap of pathway Combined scores}
\usage{
path_heatmap(
  df_list,
  libs = "default",
  analysis = c("PFG", "EnrichR"),
  legend_title = "Combined\\nScore",
  legend_direction = "vertical",
  cluster_columns = TRUE,
  anno_row = TRUE,
  colors_ht = NULL,
  plot_title = NULL,
  row_fontsize = 10,
  col_fontsize = 10,
  title_fontsize = 12,
  max_str_width = 40,
  n_legend_breaks = 5,
  n_color_breaks = 11,
  score_cutoff = 1,
  top_n_paths = 20,
  colors_anno_row = c(pal_jco()(2)[2:1], "red2", brewer.pal(7, "Dark2")[c(7, 1, 2:6)])
)
}
\description{
Function to plot Heatmap of pathway Combined scores
}
\examples{
## Placeholder Example ##
}
