% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_bars}
\alias{plot_bars}
\title{Function to create barplots}
\usage{
plot_bars(
  df,
  x_var,
  y_var,
  fill_var,
  x_angle = 45,
  legend_name = NULL,
  title = NULL,
  y_lab = NULL,
  x_lab = NULL,
  barcolors = NULL,
  legend_position = "bottom",
  bar_type = c("dodge", "stack")
)
}
\description{
Function to create barplots
}
\examples{
## Placeholder Example ##
}
