% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_corr}
\alias{plot_corr}
\title{Function to generate correlation plots}
\usage{
plot_corr(
  df,
  proteins_of_interest,
  proteins_of_interest_col = "query_protein",
  partner_proteins_col = "protein",
  corr_col = "corr",
  is_significant_col = "is_signif",
  qvalue_col = "qval",
  only_significant = FALSE,
  display_mode = c("lower_triangle", "upper_triangle", "full"),
  show_legend = TRUE,
  x_axis_text_angle = 45,
  n_panels = 1,
  plot_title = NULL,
  na_color = "grey90",
  corr_colors = colorRampPalette(c("#053061", "#2166AC", "#4393C3", "#92C5DE", "#D1E5F0",
    "#FFFFFF", "#FDDBC7", "#F4A582", "#D6604D", "#B2182B", "#67001F"))(200)
)
}
\description{
Function to generate correlation plots
}
\examples{
## Placeholder Example ##
}
