% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_corr_heatmap}
\alias{plot_corr_heatmap}
\title{Function to generate correlation change heatmap by protein}
\usage{
plot_corr_heatmap(
  comb_corr,
  subsets,
  dataset_names,
  title_suffix = NULL,
  top_n = 50,
  cluster_rows = TRUE,
  signif_mode = c("any", "all"),
  annotate_signif = TRUE,
  signif_symbol = "✱",
  subtitle = TRUE,
  switch_mode = c("none", "dissimilar", "prioritize_dissimilar", "similar",
    "prioritize_similar")
)
}
\description{
Function to generate correlation change heatmap by protein
}
\examples{
## Placeholder Example ##
}
