% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_corr_scatter}
\alias{plot_corr_scatter}
\title{Function to generate scatter plots}
\usage{
plot_corr_scatter(
  comb_corr,
  datasetA,
  datasetB,
  datasetA_name,
  datasetB_name,
  query_protein,
  highlight_n = 20,
  colors = c("red2", "blue2", "purple2", "grey50")
)
}
\description{
Function to generate scatter plots
}
\examples{
## Placeholder Example ##
}
