% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_delta_scatter}
\alias{plot_delta_scatter}
\title{Function togenerate delta-delta scatter plots}
\usage{
plot_delta_scatter(
  comb_corr,
  datasetA,
  datasetB,
  baseline_dataset,
  query_protein,
  datasetA_name,
  datasetB_name,
  colors = c("red2", "blue2", "purple2", "grey50"),
  baseline_name,
  highlight_n = 20
)
}
\description{
Function togenerate delta-delta scatter plots
}
\examples{
## Placeholder Example ##
}
