% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{plot_delta_volcano}
\alias{plot_delta_volcano}
\title{Function to generate delta-volcano plots}
\usage{
plot_delta_volcano(
  comb_corr,
  datasetA,
  datasetB,
  datasetA_name,
  datasetB_name,
  highlight_n = 15
)
}
\description{
Function to generate delta-volcano plots
}
\examples{
## Placeholder Example ##
}
