% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_density}
\alias{plot_density}
\title{Function to make density plots}
\usage{
plot_density(
  df,
  protein,
  grouping_var = NULL,
  line_var_label = "Type",
  df_sample_type = "AML",
  title_suffix = "AML",
  df_normal = NULL,
  normal_sample_type = "CD34+",
  covar = NULL,
  covar_label = NULL,
  facet = FALSE,
  n_panels = 1,
  fill_colors = NULL,
  color_fill_mode = c("list", "vector"),
  line_colors = RColorBrewer::brewer.pal(7, "Dark2"),
  normal_fill_colors = "grey80",
  default_density_color = "steelblue",
  legend_mode = c("inside", "none", "right", "left", "bottom", "top"),
  legend_position = c(0.75, 0.75)
)
}
\description{
Function to make density plots
}
\examples{
## Placeholder Example ##
}
