% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{plot_enrichment_paths}
\alias{plot_enrichment_paths}
\title{Function to generate Enrichment Barplots}
\usage{
plot_enrichment_paths(df, type = c("PFG", "EnrichR"), path_number = 20)
}
\description{
Function to generate Enrichment Barplots
}
\examples{
## Placeholder Example ##
}
