% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_km}
\alias{plot_km}
\title{Function to make KM plots}
\usage{
plot_km(
  data,
  time_var,
  event_var,
  auto_style = FALSE,
  grouping_var = NULL,
  covar = NULL,
  title = NULL,
  ylab = "Cumulative probability",
  xlab = "Time (years)",
  xlim = NULL,
  break_x_by = NULL,
  table_by_time = FALSE,
  table_times = NULL,
  legend_mode = c("inside", "none", "right", "left", "bottom", "top"),
  legend_position = c(0.7, 0.7),
  legend_labs = NULL,
  ncol_legend = 1,
  pval_coord = c(3.5, 1),
  risk_table = FALSE,
  single_color = NULL,
  show_pval_table = TRUE,
  pval_table_adj_method = "BH",
  n_panels = 1,
  add_overall = TRUE,
  linetype_overall = "solid",
  color_overall = "black",
  line_types = c("solid", "twodash", "dashed", "dotted", "dotdash", "longdash"),
  colors = c("red2", "blue3", "green4", "purple2", "darkorange", "darkcyan",
    "darkgoldenrod3", "yellow2", "deeppink2", "lightsalmon2", "darkturquoise")
)
}
\description{
Function to make KM plots
}
\examples{
## Placeholder Example ##
}
