% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_upset}
\alias{plot_upset}
\title{Function to make upset plots}
\usage{
plot_upset(
  df_list,
  groups = NULL,
  colors = NULL,
  only_color = "grey70",
  exclusive_color = "gold",
  highlight_exclusive = FALSE,
  plot_title = NULL
)
}
\description{
Function to make upset plots
}
\examples{
## Placeholder Example ##
}
