% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{plot_vol_de}
\alias{plot_vol_de}
\title{Function to generate Volcano plot from DE analysis with limma}
\usage{
plot_vol_de(
  df,
  LFC,
  pval,
  groups,
  down_ptn,
  up_ptn,
  pval_cutoff = 0.05,
  LFC_cutoff = 0.25
)
}
\description{
Function to generate Volcano plot from DE analysis with limma
}
\examples{
## Placeholder Example ##
}
