% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{prepare_submission}
\alias{prepare_submission}
\title{Function to prepare submission folder}
\usage{
prepare_submission(
  manuscript_path,
  lookup_table,
  output_path,
  sup_naming = "Supplemental"
)
}
\description{
Function to prepare submission folder
}
\examples{
## Placeholder Example ##
}
