% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_bridging.R
\name{preprocess_corr_sig}
\alias{preprocess_corr_sig}
\title{Function to pre-process Correlations output for pathways and protein networks input}
\usage{
preprocess_corr_sig(cor_df, datasets, df_ptn_map, query_ptn = NULL)
}
\description{
Function to pre-process Correlations output for pathways and protein networks input
}
\examples{
## Placeholder Example ##
}
