% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing_wrapper.R
\name{process_references}
\alias{process_references}
\title{Function to wrap Zotero References pipeline}
\usage{
process_references(
  text,
  collection_name,
  user_id,
  api_key,
  main_cache_path,
  zotero_version = "3",
  crossref_email = NULL
)
}
\description{
Function to wrap Zotero References pipeline
}
\examples{
## Placeholder Example ##
}
