% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_data.R
\name{prog_relevel}
\alias{prog_relevel}
\title{Function to relevel grouping variable according to prognnosis}
\usage{
prog_relevel(df, group_var, time_var = "surv_time", event_var = "status")
}
\description{
Function to relevel grouping variable according to prognnosis
}
\examples{
## Placeholder Example ##
}
