% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_data.R
\name{progeny_clusters}
\alias{progeny_clusters}
\title{Function to generate clusters with progenyClust}
\usage{
progeny_clusters(
  df,
  proteins,
  n_clusters,
  cluster_prefix = "C",
  ncluster_pgcl = 2:10,
  size_pgcl = 10,
  iteration_pgcl = 100,
  repeats_pgcl = 10,
  nrandom_pgcl = 10,
  method_pgcl = "gap",
  plot_gap = TRUE,
  time_var = "surv_time",
  event_var = "status",
  relapse_var = "relapse"
)
}
\description{
Function to generate clusters with progenyClust
}
\examples{
## Placeholder Example ##
}
