% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{ptn_corr}
\alias{ptn_corr}
\title{Function to perform correlation analysis}
\usage{
ptn_corr(
  df,
  all_proteins,
  proteins_of_interest,
  method = "spearman",
  conf_level = 0.95,
  pval_cutoff = 0.05,
  abs_cutoff = 0.2,
  p_adjust_method = "BH"
)
}
\description{
Function to perform correlation analysis
}
\examples{
## Placeholder Example ##
}
