% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{qc_feature}
\alias{qc_feature}
\title{Function to check for protein (feature) outliers}
\usage{
qc_feature(
  df_sample_filtered,
  proteins,
  check_cor = TRUE,
  check_mad = FALSE,
  cor_cutoff = 0.95,
  ptn_z_thrs = 3,
  freq_thrs = 0.1
)
}
\description{
Function to check for protein (feature) outliers
}
\examples{
## Placeholder Example ##
}
