% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_analysis.R
\name{qc_sample}
\alias{qc_sample}
\title{Function to check for sample outliers}
\usage{
qc_sample(df, proteins)
}
\description{
Function to check for sample outliers
}
\examples{
## Placeholder Example ##
}
