% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{rare_sparse_vars}
\alias{rare_sparse_vars}
\title{Function to detect rare and sparse variables for removal before making descriptive tables}
\usage{
rare_sparse_vars(
  df,
  vars,
  group_var,
  mode = c("table", "outcome"),
  min_n = 5,
  max_levels = 6,
  min_freq = 0.03,
  min_epv = 10
)
}
\description{
Function to detect rare and sparse variables for removal before making descriptive tables
}
\examples{
## Placeholder Example ##
}
