% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{remove_extra_terms}
\alias{remove_extra_terms}
\title{Function to remove extra proteins that are not in working dataset}
\usage{
remove_extra_terms(data, features, reference_features, verbose = TRUE)
}
\description{
Function to remove extra proteins that are not in working dataset
}
\examples{
## Placeholder Example ##
}
