% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{render_manuscript}
\alias{render_manuscript}
\title{Function to render manuscript file}
\usage{
render_manuscript(
  manuscript_path,
  manuscript_name = "manuscript",
  mode = c("draft", "review", "submission")
)
}
\description{
Function to render manuscript file
}
\examples{
## Placeholder Example ##
}
