% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{rewrite_citations}
\alias{rewrite_citations}
\title{Function to replace text with Zotero citation keys (silently skips already converted citations)}
\usage{
rewrite_citations(text, table, citekeys)
}
\description{
Function to replace text with Zotero citation keys (silently skips already converted citations)
}
\examples{
## Placeholder Example ##
}
