% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{rewrite_figure_table_keys}
\alias{rewrite_figure_table_keys}
\title{Wrapper funtion to raw text with all Figures and Tables Quarto keys}
\usage{
rewrite_figure_table_keys(
  text,
  mfig_tag = "FIGURE MAIN",
  sfig_tag = "FIGURE SUP",
  mtbl_tag = "TABLE MAIN",
  stbl_tag = "TABLE SUP"
)
}
\description{
Wrapper funtion to raw text with all Figures and Tables Quarto keys
}
\examples{
## Placeholder Example ##
}
