% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_managing.R
\name{rewrite_tag}
\alias{rewrite_tag}
\title{Function to replace text with Quarto keys for Figures and Tables}
\usage{
rewrite_tag(text, tag, prefix)
}
\description{
Function to replace text with Quarto keys for Figures and Tables
}
\examples{
## Placeholder Example ##
}
