% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{run_bxplots}
\alias{run_bxplots}
\title{Function to wrap generation of multiple boxplots}
\usage{
run_bxplots(
  df,
  continuous_vars,
  discrete_vars,
  fill_var = NULL,
  padj = "BH",
  colors = NULL,
  disc_var_labels = NULL,
  show_pairwise = FALSE,
  title_suffix = NULL,
  angle_x_var = 0,
  label_y_test_pos_pct = 0.1,
  show_median_line = TRUE,
  legend_bxplt = FALSE,
  legend_jitter = FALSE,
  fill_label = NULL,
  ylab = NULL,
  xlab = NULL
)
}
\description{
Function to wrap generation of multiple boxplots
}
\examples{
## Placeholder Example ##
}
