% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_wrapper.R
\name{run_corr_pipeline}
\alias{run_corr_pipeline}
\title{Function to wrap correlations pipeline for multiple datasets}
\usage{
run_corr_pipeline(
  df_input,
  proteins_of_interest,
  all_proteins,
  grouping_variable = NULL,
  var_labs = NULL,
  df_normal = NULL,
  df_subsetting = FALSE,
  min_samples = 10,
  ncol_grid = 2,
  verbose = TRUE,
  method = "spearman",
  conf_level = 0.95,
  pval_cutoff = 0.05,
  abs_cutoff = 0.2,
  run_sample_qc = TRUE,
  sample_outlier_policy = "keep_all",
  sample_rm_thr = 30,
  sample_diff_thr = 20,
  run_feature_qc = TRUE,
  feature_outlier_policy = "keep_all",
  check_cor = TRUE,
  check_mad = FALSE,
  cor_cutoff = 0.95,
  ptn_z_thrs = 3,
  freq_thrs = 0.1,
  p_adjust_method = "BH",
  single_dataset_label = "all"
)
}
\description{
Function to wrap correlations pipeline for multiple datasets
}
\examples{
## Placeholder Example ##
}
