% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{run_coxph}
\alias{run_coxph}
\title{Function to compute Univariate CoxPH}
\usage{
run_coxph(data, features, time_var, event_var, feat_labs, verbose = TRUE)
}
\description{
Function to compute Univariate CoxPH
}
\examples{
## Placeholder Example ##
}
