% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_wrapper.R
\name{run_DE_analysis}
\alias{run_DE_analysis}
\title{Function to wrap DE analysis pipeline}
\usage{
run_DE_analysis(
  df,
  proteins,
  grouping_variable = NULL,
  verbose = TRUE,
  run_sample_qc = TRUE,
  sample_outlier_policy = "keep_all",
  sample_rm_thr = 30,
  sample_diff_thr = 20,
  run_feature_qc = TRUE,
  feature_outlier_policy = "keep_all",
  check_cor = TRUE,
  check_mad = FALSE,
  cor_cutoff = 0.95,
  ptn_z_thrs = 3,
  freq_thrs = 0.1,
  de_mode = c("binary", "ordered"),
  pval_cutoff = 0.05,
  LFC_cutoff = 0.25,
  ref_level = NULL,
  strict = FALSE,
  treat = FALSE,
  covars = NULL,
  vplot_legend_position = c(0.8, 0.9)
)
}
\description{
Function to wrap DE analysis pipeline
}
\examples{
## Placeholder Example ##
}
