% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_wrapper.R
\name{run_enriched_paths}
\alias{run_enriched_paths}
\title{Function to wrap Pathway analysis pipeline}
\usage{
run_enriched_paths(
  proteins,
  analysis = c("all", "PFG", "EnrichR"),
  original_name_key = "original",
  pfg_list = NULL,
  repository_name_key = "repository",
  libs = "default",
  enrichr_filt_ptn = FALSE,
  min_ptn_enrichr_path = 3,
  enrichr_filt_path = FALSE,
  min_path_enrichr_ptn = 3,
  out_rm = TRUE,
  drop_paths = TRUE,
  path_number = 20,
  max_str_width = 40
)
}
\description{
Function to wrap Pathway analysis pipeline
}
\examples{
## Placeholder Example ##
}
