% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{run_mv_coxph}
\alias{run_mv_coxph}
\title{Function to generate stepwise forward MV models}
\usage{
run_mv_coxph(
  df,
  meta_table,
  vars,
  time_var,
  event_var,
  grouping_variable,
  priority_levels = c("core", "optional", "low"),
  selection_strategy = c("strict", "two_phase", "cie"),
  cie_threshold = 0.1,
  cie_final_cleanup = FALSE,
  priority_override = NULL,
  min_epv = 10,
  max_vars = NULL,
  cluster_var_p_cutoff = 0.05,
  labels = NULL,
  max_missing_mv = 0.2,
  warn_missing_n = 50,
  lrt_p_cutoff = 0.05,
  block_vars = TRUE,
  force_mv_vars = NULL,
  all_levels_sig = FALSE,
  run_bootstrap = FALSE,
  n_bootstrap = 200,
  bootstrap_cutoff = 0.5,
  parallel = FALSE,
  seed = 123,
  n_cores = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\description{
Function to generate stepwise forward MV models
}
\examples{
## Placeholder Example ##
}
