% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{run_mv_sensitivity}
\alias{run_mv_sensitivity}
\title{Function to define best MVA strategy}
\usage{
run_mv_sensitivity(
  df,
  meta_table,
  time_var,
  event_var,
  vars = NULL,
  grouping_variable,
  group_var_label = "Group",
  strategies = c("strict", "two_phase", "cie"),
  cie_threshold = 0.1,
  cie_final_cleanup = FALSE,
  priority_levels = c("core", "optional", "low"),
  priority_override = NULL,
  block_vars = TRUE,
  min_epv = 10,
  max_vars = NULL,
  max_missing_mv = 0.2,
  warn_missing_n = 50,
  cluster_var_p_cutoff = 0.05,
  lrt_p_cutoff = 0.05,
  label_vars = NULL,
  relevel_vars = NULL,
  dichotomous_vars = NULL,
  force_mv_vars = NULL,
  mv_independent = TRUE,
  run_bootstrap = FALSE,
  n_bootstrap = 200,
  bootstrap_cutoff = 0.5,
  parallel = FALSE,
  seed = 123,
  n_cores = parallel::detectCores() - 1,
  verbose = TRUE,
  all_levels_sig = FALSE
)
}
\description{
Function to define best MVA strategy
}
\examples{
## Placeholder Example ##
}
