% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{run_nested_cv}
\alias{run_nested_cv}
\title{Function to perform nested CV on train data with optional parallelization}
\usage{
run_nested_cv(
  data,
  features,
  time_var,
  event_var,
  id_col = NULL,
  lambda = c("lambda.min", "lambda.1se"),
  alpha_grid = seq(0.1, 1, by = 0.1),
  n_outer = 5,
  n_inner = 5,
  seed = 42,
  verbose = TRUE,
  n_cores = 1,
  parallelize = TRUE
)
}
\description{
Function to perform nested CV on train data with optional parallelization
}
\examples{
## Placeholder Example ##
}
