% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{run_outer_fold}
\alias{run_outer_fold}
\title{Function to run one outer fold in nested CV}
\usage{
run_outer_fold(
  split,
  alpha_grid,
  n_inner,
  features,
  time_var,
  event_var,
  lambda = c("lambda.min", "lambda.1se")
)
}
\description{
Function to run one outer fold in nested CV
}
\examples{
## Placeholder Example ##
}
