% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{run_panel_reduction}
\alias{run_panel_reduction}
\title{Function to rank features and run panel reduction}
\usage{
run_panel_reduction(
  data,
  features,
  stab_freq,
  time_var,
  event_var,
  alpha,
  coefs,
  lambda,
  min_size = 5,
  tolerance = 0.02,
  n_outer = 5,
  n_inner = 5,
  seed = 42,
  id_col = NULL,
  verbose = TRUE,
  n_cores = 1,
  parallelize = TRUE
)
}
\description{
Function to rank features and run panel reduction
}
\examples{
## Placeholder Example ##
}
