% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{run_stability_selection}
\alias{run_stability_selection}
\title{Function to perform feature stability selection on train data with optional parallelization}
\usage{
run_stability_selection(
  data,
  features,
  time_var,
  event_var,
  alpha,
  lambda,
  stab_cutoff = 0.75,
  b = 100,
  seed = 42,
  n_cores = 1,
  parallelize = TRUE,
  verbose = TRUE
)
}
\description{
Function to perform feature stability selection on train data with optional parallelization
}
\examples{
## Placeholder Example ##
}
