% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_tables_uv_mv_analysis.R
\name{run_uv_coxph}
\alias{run_uv_coxph}
\title{Function to compute UV models and extract best variables for MVA}
\usage{
run_uv_coxph(
  df,
  vars,
  time_var,
  event_var,
  uv_p_cutoff = 0.2,
  labels = NULL,
  meta_table = NULL,
  group_var_label = NULL
)
}
\description{
Function to compute UV models and extract best variables for MVA
}
\examples{
## Placeholder Example ##
}
