% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_data.R
\name{safe_clean_raw_data}
\alias{safe_clean_raw_data}
\title{Function to safe clean raw data}
\usage{
safe_clean_raw_data(
  data,
  corrections_table,
  numeric_vars = NULL,
  na_strings = c("na", "n/a", "n.a", "nan", "null", "none", "missing", "not available",
    "not applicable", "unknown", "", "--", "---")
)
}
\description{
Function to safe clean raw data
}
\examples{
## Placeholder Example ##
}
