% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_networks_wrapper.R
\name{set_contrast}
\alias{set_contrast}
\title{Function to set constrasts within the protein networks pipeline}
\usage{
set_contrast(
  contrast,
  filename,
  edge_palette = NULL,
  wait_for_legend = TRUE,
  expr_col = NULL,
  width_col = NULL,
  class_col = NULL
)
}
\description{
Function to set constrasts within the protein networks pipeline
}
\examples{
## Placeholder Example ##
}
