% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_utilities.R
\name{sync_csl_files}
\alias{sync_csl_files}
\title{Function to sync CSL entries}
\usage{
sync_csl_files(csl_dir, registry)
}
\description{
Function to sync CSL entries
}
\examples{
## Placeholder Example ##
}
