% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{validate_doi}
\alias{validate_doi}
\title{Function to validate DOIs}
\usage{
validate_doi(doi, crossref_email = NULL)
}
\description{
Function to validate DOIs
}
\examples{
## Placeholder Example ##
}
