% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{wf_plot}
\alias{wf_plot}
\title{Function to generate Waterfall plots}
\usage{
wf_plot(df, type = c("Positive", "Negative"), abs_cutoff = 0.2)
}
\description{
Function to generate Waterfall plots
}
\examples{
## Placeholder Example ##
}
