% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_managing.R
\name{zotero_headers}
\alias{zotero_headers}
\title{Function to pass zotero headers}
\usage{
zotero_headers(api_key, zotero_version = "3")
}
\description{
Function to pass zotero headers
}
\examples{
## Placeholder Example ##
}
