% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Function to make heatmaps}
\usage{
plot_heatmap(
  df,
  proteins,
  col_anno = TRUE,
  grouping_var = NULL,
  covars = NULL,
  covar_labels = NULL,
  grouping_var_label = NULL,
  row_anno = FALSE,
  row_anno_df = NULL,
  row_anno_protein_col = NULL,
  rown_anno_vars = NULL,
  rown_anno_vars_labels = NULL,
  colors_ht = NULL,
  covar_colors_override = NULL,
  colors_anno_col = c(pal_jco()(2)[2:1], "red2", brewer.pal(7, "Dark2")[c(7, 1, 2:6)]),
  colors_anno_row = c(pal_jco()(2)[2:1], "red2", brewer.pal(7, "Dark2")[c(7, 1, 2:6)]),
  colors_grouping_var = brewer.pal(7, "Dark2"),
  colors_NA = pal_jco()(3)[3],
  plot_title = NULL,
  cluster_rows = TRUE,
  cluster_cols = FALSE,
  legend_breaks = NULL,
  legend_labels = NULL,
  breaks = NULL,
  protein_labels = NULL,
  show_colnames = FALSE,
  show_rownames = TRUE,
  annotation_names_row = FALSE
)
}
\description{
Function to make heatmaps
}
\examples{
## Placeholder Example ##
}
