% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_data_wrapper.R
\name{run_QC_analysis}
\alias{run_QC_analysis}
\title{Function to wrap QC pipeline}
\usage{
run_QC_analysis(
  df,
  proteins,
  verbose = TRUE,
  grouping_variable = NULL,
  run_sample_qc = TRUE,
  sample_outlier_policy = "keep_all",
  sample_rm_thr = 30,
  sample_diff_thr = 20,
  run_feature_qc = TRUE,
  feature_outlier_policy = "keep_all",
  check_cor = TRUE,
  check_mad = FALSE,
  cor_cutoff = 0.95,
  ptn_z_thrs = 3,
  freq_thrs = 0.1
)
}
\description{
Function to wrap QC pipeline
}
\examples{
## Placeholder Example ##
}
